Last step of #19 - file reorg. Fixing double free. Making free_map idempotent
