+++ includes/cub3D.h
@@ -93,6 +93,7 @@ int				player_check(t_map *map, int map_start);
 void			render_view(t_game *game);
 void			setup_input(t_game *game);
 void	draw_background(t_game *game);
+void game_loop(t_game *game);
 
 
 #endif
